requirements.txt
# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
optuna>=3.4.0
pandas>=2.1.0
numpy>=1.25.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# APIs
requests>=2.31.0

# Utilities
scipy>=1.11.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Development / Notebooks
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0